Fix 1 issue with newListenerEffect:
If a component only wishes to use one or several of the store's *actions*,
but is not going to use any part of the *state*, this component should
*not* be re-rendered when the state changes.

This should be possible by providing a `mapState` function which return
value would always match `newListener.oldState`.

But `oldState` is initialized with an anonymous object reference
(`oldState: {}`) which can not be compared against.

Providing a `mapState` function of:

`(state) => ({})`

would not work as the two references are not equal, and the test on line
13 (`if (mappedState === newListener.oldState)`) would fail.

For this test to succeed, and therefore for actions-only components not
to be re-rendered on each state change, the initial `oldState` must be
equal to each return value from `mapState`. This commit changes the
initial value of `oldState` to `undefined`, so that any component can
write something like :

`const [, { actionOne, actionTwo }] = useXyz(() => undefined);`

and never be re-rendered on state changes, while still having access to
actions.